lang/claude-code-runtime-python/pyproject.toml
[project]
name = "claude-code-runtime"
version = "0.1.0"
description = "EAASP L1 Runtime — Python T1 Harness wrapping claude-agent-sdk"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "grpcio>=1.70.0",
    "grpcio-tools>=1.70.0",
    "protobuf>=5.0",
    "anyio>=4.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_code_runtime"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
